+++ src/scene.c
@@ -7,7 +7,7 @@
 #include "bvh.h"
 #include "tlas.h"
 
-void scene_init(scene *s, size_t mesh_cnt, size_t inst_cnt, size_t mat_cnt)
+void scene_init(scene *s, uint32_t mesh_cnt, uint32_t inst_cnt, uint32_t mat_cnt)
 {
   s->mesh_cnt = mesh_cnt;
   s->inst_cnt = inst_cnt;
